pyproject.toml
[tool.poetry]
name = "repl_python3_beta_Minecraft-Status-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
mcstatus = "^5.2.0"
python-decouple = "^3.6"
Flask = "^2.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
